requirements.txt
Django==3.2.7
selenium==3.141.0
psycopg2-binary==2.8.6
python-dotenv==1.0.0
Telethon==1.28.5